django_project/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1
django-crispy-forms==1.7.2
idna==2.7
Pillow==5.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
markdown==3.7
django-markdownx==4.0.7